pyproject.toml
[project]
authors         = [{ name = "Viktor Vasylkovskyi", email = "viktor.vasylkovskyi@gmail.com" }]
description     = "A most simple LLM that barks and responds to user commands."
license         = { text = "MIT" }                                                            # adjust if different
maintainers     = [{ name = "Viktor Vasylkovskyi", email = "viktor.vasylkovskyi@gmail.com" }]
name            = "bark-gpt"
readme          = "README.md"
requires-python = ">=3.12"
version         = "0.0.1"

dependencies = [
  "a2a-sdk>=0.3.22",
  "aiokafka==0.10.0",
  "colorama==0.4.6",
  "fastapi>=0.124.4",
  "pydantic (>=2.11.1,<3.0.0)",
  "pydantic-settings (>=2.8.1,<3.0.0)",
  "torch>=2.9.1",
  "transformers>=4.57.3",
  "uvicorn>=0.38.0",
]

[project.urls]
Homepage   = "https://github.com/vvasylkovskyi/bark-gpt"
Repository = "https://github.com/vvasylkovskyi/bark-gpt"

[tool.poetry.group.dev.dependencies]
dependencies = ["pyright==1.1.374", "ruff==0.5.5"]

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]
fix = true
line-length = 120
src = ["src", "tests"]
target-version = "py311"

[tool.ruff.format]
indent-style              = "space"
line-ending               = "lf"
quote-style               = "double"
skip-magic-trailing-comma = true

[tool.pyright]
exclude                        = ["**/__pycache__"]
ignore                         = []
include                        = ["src"]
reportMissingImports           = true
reportMissingTypeStubs         = false
reportUnknownParameterType     = true
reportUntypedFunctionDecorator = true
strict                         = []
venvPath                       = ".venv"
